[Complete] User Module & Basic Functionality

+++ src/app/module/user/user.services.ts
@@ -1,6 +1,11 @@
+import paginationHelper from "../../../helpers/paginationHelper";
 import { APIError } from "../../../interface/APIError";
-import { IUser } from "./user.interface";
+import { IPagination } from "../../../interface/Pagination";
+import { IAllDataReturnType } from "../../../interface/common";
+import { userSearchFields } from "./user.constant";
+import { IUser, IUserSearchFilter } from "./user.interface";
 import User from "./user.model";
+import isUserFound from "./user.utils";
 
 const signup = async (payload: IUser): Promise<IUser | null> => {
   // Business Logic: Buyer -> budgaet < 20k
@@ -18,12 +23,54 @@ const signup = async (payload: IUser): Promise<IUser | null> => {
   return data;
 };
 
-const getAllUsers = async (): Promise<IUser[] | null> => {
-  const data = await User.find();
-  return data;
+const getAllUsers = async (
+  paginationOptions: IPagination,
+  searchFilterFields: IUserSearchFilter
+): Promise<IAllDataReturnType<IUser[]> | null> => {
+  // Pagination
+  const { page, limit, skip, sortBy, sortOrder } =
+    paginationHelper(paginationOptions);
+
+  // Sort Condition
+  const sortCondition = { [sortBy]: sortOrder };
+
+  const { searchTerm, ...filterdata } = searchFilterFields;
+  const andCondition = [];
+
+  // Search Condition
+  if (searchTerm) {
+    andCondition.push({
+      $or: userSearchFields.map((field) => ({
+        [field]: { $regex: searchTerm, $options: "i" },
+      })),
+    });
+  }
+
+  // Filter Fields
+  if (Object.keys(filterdata).length) {
+    andCondition.push({
+      $and: Object.entries(filterdata).map(([field, value]) => ({
+        [field]: [value],
+      })),
+    });
+  }
+
+  const whereCondition = andCondition.length ? { $and: andCondition } : {};
+
+  const data = await User.find(whereCondition)
+    .sort(sortCondition)
+    .skip(skip)
+    .limit(limit);
+  const total = await User.countDocuments();
+  const meta = { page, limit, total };
+  return { meta, data };
 };
 
 const getUser = async (id: string): Promise<IUser | null> => {
+  if (!(await isUserFound(id))) {
+    throw new APIError(400, "User not Found!");
+  }
+
   const data = await User.findById(id);
   return data;
 };
@@ -32,7 +79,20 @@ const updateUser = async (
   _id: string,
   payload: IUser
 ): Promise<IUser | null> => {
-  const data = await User.findOneAndUpdate({ _id }, payload, {
+  if (!(await isUserFound(_id))) {
+    throw new APIError(400, "User not Found!");
+  }
+
+  let { name, ...userData } = payload;
+
+  if (name && Object.keys(name).length) {
+    Object.keys(name).map((field) => {
+      const nameKey = `name.${field}`;
+      (userData as any)[nameKey] = name[field as keyof typeof name];
+    });
+  }
+
+  const data = await User.findOneAndUpdate({ _id }, userData, {
     new: true,
     runValidators: true,
   });
@@ -41,6 +101,10 @@ const updateUser = async (
 };
 
 const deleteUser = async (id: string): Promise<IUser | null> => {
+  if (!(await isUserFound(id))) {
+    throw new APIError(400, "User not Found!");
+  }
+
   const data = await User.findByIdAndDelete(id);
   return data;
 };
